Rename getmeals.js to apiPosts.js and added likes api url
