+++ src/modules/consts.js
@@ -1,2 +1,3 @@
 const API_URL = 'https://www.themealdb.com/api/json/v1/1/filter.php?c=Seafood';
-export default API_URL;
+const LIKES_URL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/vehS1XVjnLmMzjCWvntO/likes';
+export {API_URL, LIKES_URL};
